pyproject.toml
[project]
name = "sitemap-links-issues"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiohttp (>=3.11.13,<4.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "black (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
